feat: Add `use_null_prototype` option where applicable.
This library defaults to use a `null` prototype when creating objects,
which is slightly more efficient. However, deep equality assertions such
as NodeJS's `deepStringEqual` will fail if the prototypes don't match.
This causes users to have to re-create certain objects when testing,
which is cumbersome.

This adds `use_null_prototype` which will use `Object` as the prototype
when objects are created. This is mostly useful in `filter`, however the
option was added across multiple functions for consistency.